src/agents/__init__.py
from .base_agent import BaseAgent
from .core.experimental_agent import ExperimentalAgent
from .core.theoretical_agent import TheoreticalAgent
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("OPENAI_API_KEY")

__all__ = ["BaseAgent", "ExperimentalAgent", "TheoreticalAgent", "CommunicationAgent"]