pyproject.toml
[tool.poetry]
name = "SimuSearch"
version = "0.1.0"
description = "A LangChain-based multi-agent research system using OpenRouter"
authors = ["Mostafa Faheem <mostafaaafaheem@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.0"
langchain-openai = "^0.1.7"
openai = "^1.42.0"
httpx = "^0.27.0"
pydantic = "^2.9.0"
tqdm = "^4.66.0"
rich = "^13.7.1"
python-dotenv = "^1.1.1"
openrouter = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
black = "^24.4.2"
ruff = "^0.5.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"